backend/requirements.txt
azure-functions
azure-storage-blob==12.19.0
azure-cosmos==4.5.1
azure-cognitiveservices-speech==1.36.0
azure-identity==1.17.0